Update 3.6.2
**Bug 2 Resolved**
Adjusted the sprite rotation speed by increasing the speed for a smoother translation. I also changed it from lerping from one point to another to smooth stepping (easing in and out from one direction to the other).

**Bug 10 & 11 Resolved**
The attack was set using the old input system. I have updated it so it now uses the new input system and I removed the function that was checking to see where the mouse cursor was and if it wasn't infront of the player it would not melee, now the player will melee everytime the press the lmb. The mouse cursor is hidden when playing and is showing when pause or in a menu screen.