Update 3.6.5
**Bug 8: Resolved**
Yarn now shoots infinitely. I have kept the logic to change this to have the need to have it consumed and replenished if we wanted to revert back to the original concept.

**Bug 15: Resolved**
The tutorial signs now display A & D to move and Space & W to jump.

**Bug 20: Resolved**
Made the hitboxes for the platforms just a little bigger to give the player a small margin of error.