Update 3.6.7
BUG 24: RESOLVED

Start button was not working because I was attempting to have the SFX for the button press to play through before the scene change. It was either not playing because the scene change was happening instantaneously even though the SFX was being called to play or it played and never changed scenes.
It was not changing scenes after adding the coroutine because It was not registering that the button pressed boolean was not false.
I implemented a reusable function/coroutine that looks for an action and fires it after playing the SFX. This corrected the issue and will now allow me to use it on any buttons that change the scene or end the application.

BUG 16: RESOLVED
When releasing the sfx slider it plays the Cat SFX. I made a helper script and attached it to the slider, it checks for an on point up when this slider is pressed.

Bug 19, 21, 22, 23, 25: Resolved
Major issues with this system. However, it was finally corrected. Master now mutes all three audio mixers, all buttons in the main menu have sfx, default button now unmute and untoggles and lastly the mute button image triggers when the volume reaches 0.