Merge branch 'Level-Design-Tutorial-Level' into Level_Select
- Installed URP (I thought this was already on)
- Added stationary platforms and adjusted their sizes and colliders.
- Added moving platforms and adjusted their sizes and colliders.
- Changed the player life UI displayer to show the cats face and x00 (similar to Mario games)
- Changed the UI for the coins and added the rotating coin instead of the static image.
- Made a cloud loop script that moves and spawns the clouds in the background.
- Fog has been added to level 1
- Bug 8: Resolved
Had the wrong platform attached to the button script. It was not telling the correct platform that was not moving that it could move.
- Bug 9: Resolved
Adjusted Fall multiplier to x1. The player's fall is a bit more floatier.
- Bug 10: Resolved
Corpse collider being created in script was not correct. Changed the offset value of the collider bound when creating the new collider to the correct value.
- Bug 3: Resolved
Added a '/' between the space bar and the w on the sign on level 1